+++ jrasp-daemon/environ/environ.go
@@ -29,6 +29,7 @@ type Environ struct {
 	HostName    string `json:"hostName"`    // 主机/容器名称
 	Ip          string `json:"ip"`          // ipAddress
 	OsType      string `json:"osType"`      // 操作系统类型
+	BinFileName string `json:"binFileName"` // 磁盘可执行文件名
 	BinFileHash string `json:"binFileHash"` // 磁盘可执行文件的md5值
 
 	// 系统信息
@@ -48,6 +49,9 @@ type Environ struct {
 	BuildDecryptKey string `json:"buildDecryptKey"`
 
 	PidFile string `json:"pidFile"`
+
+	// 是否已连接server
+	IsConnectServer bool `json:"isConnectServer"`
 }
 
 func NewEnviron() (*Environ, error) {
@@ -65,6 +69,8 @@ func NewEnviron() (*Environ, error) {
 	if err != nil {
 		return nil, err
 	}
+	// 可执行文件名
+	execName := filepath.Base(execPath)
 	// ip
 	ipAddress, err := GetDefaultIp()
 	if err != nil {
@@ -86,6 +92,7 @@ func NewEnviron() (*Environ, error) {
 		InstallDir:      execDir,
 		OsType:          runtime.GOOS,
 		BinFileHash:     md5Str,
+		BinFileName:     execName,
 		TotalMem:        memInfo.Total / GB,
 		CpuCounts:       cpuCounts,
 		FreeDisk:        FreeDisk,
@@ -95,6 +102,7 @@ func NewEnviron() (*Environ, error) {
 		BuildDecryptKey: BuildDecryptKey,
 		BuildGitCommit:  BuildGitCommit,
 		PidFile:         filepath.Join(execDir, defs.DAEMON_PID_FILE),
+		IsConnectServer: false,
 	}
 	return env, nil
 }
@@ -150,6 +158,30 @@ func GetDefaultIp() (string, error) {
 	localAddr := conn.LocalAddr().(*net.UDPAddr)
 	ip := strings.Split(localAddr.IP.String(), ":")[0]
 	return ip, nil
+}
+
+func GetDefaultIface() (*net.Interface, error) {
+	defaultIP, err := GetDefaultIp()
+	if err != nil {
+		return nil, err
+	}
+	ifaces, err := net.Interfaces()
+	if err != nil {
+		return nil, err
+	}
+	for _, item := range ifaces {
+		addrs, err := item.Addrs()
+		if err != nil {
+			return nil, err
+		}
+		for _, addr := range addrs {
+			ip := addr.(*net.IPNet)
+			if ip.IP.String() == defaultIP {
+				return &item, nil
+			}
+		}
+	}
+	return nil, errors.New("not found default ifaces")
 }
 
 func GetInstallDisk(path string) (free uint64, err error) {